Исправил QuestionViewModel добавил интенты и юз кейсы, осталось сделать вью модель для предмета

+++ app/src/main/java/com/example/learningapp/domain/usecase/association/AddAssociationUseCase.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class AddAssociationUseCase @Inject constructor(
     private val repository: QuestionRepository
 ){
-    suspend operator fun invoke(association: Association) : Int{
+    suspend operator fun invoke(association: Association) : Long{
         return repository.addAssociation(association)
     }
 }
