+++ app/src/main/java/com/example/learningapp/domain/usecase/subject/AddSubjectUseCase.kt
@@ -8,7 +8,7 @@ class AddSubjectUseCase @Inject constructor(
     private val repository: QuestionRepository
 ) {
 
-    suspend operator fun invoke(subject: Subject) : Int{
+    suspend operator fun invoke(subject: Subject) : Long{
         return repository.addSubject(subject)
     }
